refactor(coach): extract account-identity from backfill
